requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
anndata==0.10.5.post1
anndata2ri==1.3.1
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
array_api_compat==1.5
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blessed==1.20.0
blinker==1.9.0
cached-property==1.5.2
cell-gears==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
cloudpickle==3.1.0
comm==0.2.1
conda-pack==0.7.1
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
dask==2024.12.1
datasets==2.14.5
dcor==0.6
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
dm-tree==0.1.8
docker-pycreds==0.4.0
docrep==0.3.2
einops==0.7.0
environs==9.5.0
et-xmlfile==1.1.0
etils==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flash-attn==1.0.6
FlashMHA==0.6.3
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
flax==0.8.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.6.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
gpustat==1.1.1
greenlet==3.1.1
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.21.4
idna==3.6
igraph==0.9.11
imbalanced-learn==0.12.3
importlib_metadata==8.5.0
importlib_resources==6.1.3
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
itsdangerous==2.2.0
jax==0.4.19
jaxlib==0.4.19
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.3.10
langchain-core==0.3.22
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langgraph==0.2.56
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.43
langsmith==0.1.147
leidenalg==0.8.10
lightning==2.1.4
lightning-utilities==0.10.1
llvmlite==0.38.1
locket==1.0.0
louvain==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
milvus-lite==2.4.10
ml-collections==0.1.1
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
mudata==0.2.3
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.15
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.55.2
numpy==1.22.4
numpyro==0.14.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.57.1
openpyxl==3.1.2
opt-einsum==3.3.0
optax==0.1.9
orbax==0.1.7
orbax-checkpoint==0.5.3
orjson==3.10.12
packaging==23.2
pandas==1.4.3
parso==0.8.3
partd==1.4.2
patsy==0.5.6
peft==0.9.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pyDeprecate==0.3.1
pydot==2.0.0
Pygments==2.17.2
pymilvus==2.4.9
pynndescent==0.5.11
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.8.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
pytorch-lightning==1.5.10
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpy2==3.5.15
safetensors==0.4.2
scanpy==1.9.8
scib==1.0.4
scikit-learn==1.4.1.post1
scikit-misc==0.1.4
scipy==1.12.0
scvi-tools==0.16.4
seaborn==0.13.2
sentry-sdk==1.41.0
session-info==1.0.0
setproctitle==1.3.3
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.12
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorstore==0.1.53
texttable==1.7.0
threadpoolctl==3.3.0
tiktoken==0.8.0
tokenizers==0.13.3
toolz==0.12.1
torch==1.13.0+cu117 --extra-index-url https://download.pytorch.org/whl/cu117
torchvision==0.14.0+cu117 --extra-index-url https://download.pytorch.org/whl/cu117
torchaudio==0.13.0+cu117 --extra-index-url https://download.pytorch.org/whl/cu117
torchdata==0.5.0
torchtext==0.14.0
torchmetrics==1.3.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.33.2
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.5
urllib3==2.2.1
wandb==0.16.4
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.21.0
